Refactored Universe.__init__ to separate topology and coordinate logic
