+++ package/MDAnalysis/coordinates/core.py
@@ -34,15 +34,9 @@ from __future__ import absolute_import
 
 import six
 
-from . import (
-    _READERS,
-    _SINGLEFRAME_WRITERS,
-    _MULTIFRAME_WRITERS,
-)
-
 from ..lib import util
-from ..lib.util import get_reader_for, get_writer_for
 from ..lib.mdamath import triclinic_box, triclinic_vectors, box_volume
+from ..core._get_readers import get_reader_for, get_writer_for
 
 
 def reader(filename, **kwargs):
