+++ package/MDAnalysis/topology/core.py
@@ -34,14 +34,13 @@ import numpy as np
 from collections import defaultdict
 
 # Local imports
-from . import _PARSERS
 from . import tables
 from .guessers import (
     guess_atom_element, guess_atom_type,
     get_atom_mass, guess_atom_mass, guess_atom_charge,
     guess_bonds, guess_angles, guess_dihedrals, guess_improper_dihedrals,
 )
-from ..lib.util import cached, get_parser_for
-from ..lib import util
+from ..core._get_readers import get_parser_for
+from ..lib.util import cached
 
 #tumbleweed
